Restructured PREQ and PREP

+++ src/devices/mesh/802.11s/mesh-configuration-element.h
@@ -106,7 +106,7 @@ protected:
   uint16_t  GetInformationSize () const;
   uint8_t GetLengthField () const;
   void SerializeInformation (Buffer::Iterator i) const;
-  uint8_t DeserializeInformation (Buffer::Iterator i, uint8_t length);
+  uint16_t DeserializeInformation (Buffer::Iterator i, uint8_t length);
   void PrintInformation(std::ostream& os) const;
   // TODO: Release and fill other fields in configuration
   // element
