+++ src/devices/mesh/802.11s/wifi-information-element.h
@@ -141,7 +141,7 @@ protected:
   /// Serialize information
   virtual void SerializeInformation (Buffer::Iterator start) const = 0;
   /// Deserialize information
-  virtual uint8_t DeserializeInformation (Buffer::Iterator start, uint8_t length) = 0;
+  virtual uint16_t DeserializeInformation (Buffer::Iterator start, uint8_t length) = 0;
   /// Print information
   virtual void PrintInformation (std::ostream &os) const = 0;
   //\}
